+++ lib/interceptor/interceptor.dart
@@ -1,0 +1,2 @@
+export 'src/gql_interceptor.dart';
+export 'src/gql_interceptor_link.dart';